splits rbac generator into two controllers
