+++ api/pkg/controller/rbac/cluster_provider.go
@@ -16,13 +16,6 @@ import (
 	"k8s.io/client-go/tools/cache"
 )
 
-const (
-	// MasterProviderPrefix denotes prefix a master cluster has in its name
-	MasterProviderPrefix = "master"
-	// SeedProviderPrefix denotes prefix a seed cluster has in its name
-	SeedProviderPrefix = "seed"
-)
-
 // ClusterProvider holds set of clients that allow for communication with the cluster and
 // that are required to properly generate RBAC for resources in that particular cluster
 type ClusterProvider struct {
