+++ api/pkg/controller/rbac/sync_project_test.go
@@ -21,7 +21,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/diff"
 	"k8s.io/client-go/kubernetes/fake"
-	rbaclister "k8s.io/client-go/listers/rbac/v1"
 	clienttesting "k8s.io/client-go/testing"
 	"k8s.io/client-go/tools/cache"
 )
@@ -54,7 +53,7 @@ func TestEnsureProjectIsInActivePhase(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			err := target.ensureProjectIsInActivePhase(test.projectToSync)
 
@@ -115,7 +114,7 @@ func TestEnsureProjectInitialized(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			err := target.ensureCleanupFinalizerExists(test.projectToSync)
 
@@ -504,7 +503,7 @@ func TestEnsureProjectClusterRBACRoleBindingForResources(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.projectResources = test.projectResourcesToSync
 			target.seedClusterProviders = seedClusterProviders
@@ -762,7 +761,7 @@ func TestEnsureClusterResourcesCleanup(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.seedClusterProviders = seedClusterProviders
 			target.userLister = userLister
 			target.masterClusterProvider = fakeMasterClusterProvider
@@ -1028,7 +1027,7 @@ func TestEnsureProjectCleanup(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.projectResources = test.projectResourcesToSync
 			target.seedClusterProviders = seedClusterProviders
@@ -1113,368 +1112,6 @@ func TestEnsureProjectCleanup(t *testing.T) {
 						t.Fatalf("%v", diff.ObjectDiff(test.expectedClusterRoleBindingsForSeeds[createActionIndex], createAction.GetObject().(*rbacv1.ClusterRoleBinding)))
 					}
 					createActionIndex = createActionIndex + 1
-				}
-			}
-		})
-	}
-}
-
-func TestEnsureProjectClusterRBACRoleBindingForNamedResource(t *testing.T) {
-	tests := []struct {
-		name                        string
-		projectToSync               *kubermaticv1.Project
-		expectedClusterRoleBindings []*rbacv1.ClusterRoleBinding
-		existingClusterRoleBindings []*rbacv1.ClusterRoleBinding
-		expectedActions             []string
-	}{
-		// scenario 1
-		{
-			name:            "scenario 1: desired RBAC Role Bindings for a project resource are created",
-			projectToSync:   test.CreateProject("thunderball", test.CreateUser("James Bond")),
-			expectedActions: []string{"create", "create", "create"},
-			expectedClusterRoleBindings: []*rbacv1.ClusterRoleBinding{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:owners-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "owners-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:owners-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "editors-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:editors-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "viewers-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:viewers-thunderball",
-					},
-				},
-			},
-		},
-
-		// scenario 2
-		{
-			name:          "scenario 2: no op when desicred RBAC Role Bindings exist",
-			projectToSync: test.CreateProject("thunderball", test.CreateUser("James Bond")),
-			existingClusterRoleBindings: []*rbacv1.ClusterRoleBinding{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:owners-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "owners-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:owners-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "editors-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:editors-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "viewers-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:viewers-thunderball",
-					},
-				},
-			},
-		},
-
-		// scenario 3
-		{
-			name:            "scenario 3: update when existing binding doesn't match desired ones",
-			projectToSync:   test.CreateProject("thunderball", test.CreateUser("James Bond")),
-			expectedActions: []string{"update", "update", "update"},
-			existingClusterRoleBindings: []*rbacv1.ClusterRoleBinding{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:owners-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "owners-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:owners-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "wrong-subject-name",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:editors-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "wrong-subject-name",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:viewers-thunderball",
-					},
-				},
-			},
-			expectedClusterRoleBindings: []*rbacv1.ClusterRoleBinding{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "editors-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:editors-thunderball",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Subjects: []rbacv1.Subject{
-						{
-							APIGroup: rbacv1.GroupName,
-							Kind:     "Group",
-							Name:     "viewers-thunderball",
-						},
-					},
-					RoleRef: rbacv1.RoleRef{
-						APIGroup: rbacv1.GroupName,
-						Kind:     "ClusterRole",
-						Name:     "kubermatic:project-thunderball:viewers-thunderball",
-					},
-				},
-			},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			// setup the test scenario
-			objs := []runtime.Object{}
-			clusterRoleBindingIndexer := cache.NewIndexer(cache.MetaNamespaceKeyFunc, cache.Indexers{})
-			for _, existingClusterRoleBinding := range test.existingClusterRoleBindings {
-				err := clusterRoleBindingIndexer.Add(existingClusterRoleBinding)
-				if err != nil {
-					t.Fatal(err)
-				}
-				objs = append(objs, existingClusterRoleBinding)
-			}
-			fakeKubeClient := fake.NewSimpleClientset(objs...)
-			clusterRoleBindingLister := rbaclister.NewClusterRoleBindingLister(clusterRoleBindingIndexer)
-
-			// act
-			target := Controller{}
-			err := target.ensureClusterRBACRoleBindingForNamedResource(test.projectToSync.Name, kubermaticv1.ProjectResourceName, kubermaticv1.ProjectKindName, test.projectToSync.GetObjectMeta(), fakeKubeClient, clusterRoleBindingLister)
-
-			// validate
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			if len(test.expectedClusterRoleBindings) == 0 {
-				if len(fakeKubeClient.Actions()) != 0 {
-					t.Fatalf("unexpected actions %#v", fakeKubeClient.Actions())
-				}
-				return
-			}
-
-			if len(fakeKubeClient.Actions()) != len(test.expectedClusterRoleBindings) {
-				t.Fatalf("unexpected actions %v", fakeKubeClient.Actions())
-			}
-			for index, action := range fakeKubeClient.Actions() {
-				if !action.Matches(test.expectedActions[index], "clusterrolebindings") {
-					t.Fatalf("unexpected action %#v", action)
-				}
-				// TODO: figure out why action.(clienttesting.GenericAction) does not work
-				createAction, ok := action.(clienttesting.CreateAction)
-				if !ok {
-					t.Fatalf("unexpected action %#v", action)
-				}
-				if !equality.Semantic.DeepEqual(createAction.GetObject().(*rbacv1.ClusterRoleBinding), test.expectedClusterRoleBindings[index]) {
-					t.Fatalf("%v", diff.ObjectDiff(test.expectedClusterRoleBindings[index], createAction.GetObject().(*rbacv1.ClusterRoleBinding)))
 				}
 			}
 		})
@@ -1649,7 +1286,7 @@ func TestEnsureProjectClusterRBACRoleForResources(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.projectResources = test.projectResourcesToSync
 			target.seedClusterProviders = seedClusterProviders
@@ -1733,308 +1370,6 @@ func TestEnsureProjectClusterRBACRoleForResources(t *testing.T) {
 	}
 }
 
-func TestEnsureProjectClusterRBACRoleForNamedResource(t *testing.T) {
-	tests := []struct {
-		name                 string
-		projectToSync        *kubermaticv1.Project
-		expectedClusterRoles []*rbacv1.ClusterRole
-		existingClusterRoles []*rbacv1.ClusterRole
-		expectedActions      []string
-	}{
-		// scenario 1
-		{
-			name:            "scenario 1: desired RBAC Roles for a project resource are created",
-			projectToSync:   test.CreateProject("thunderball", test.CreateUser("James Bond")),
-			expectedActions: []string{"create", "create", "create"},
-			expectedClusterRoles: []*rbacv1.ClusterRole{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:owners-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update", "delete"},
-						},
-					},
-				},
-
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update"},
-						},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get"},
-						},
-					},
-				},
-			},
-		},
-
-		// scenario 2
-		{
-			name:          "scenario 2: no op when desicred RBAC Roles exist",
-			projectToSync: test.CreateProject("thunderball", test.CreateUser("James Bond")),
-			existingClusterRoles: []*rbacv1.ClusterRole{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:owners-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update", "delete"},
-						},
-					},
-				},
-
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update"},
-						},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get"},
-						},
-					},
-				},
-			},
-		},
-
-		// scenario 3
-		{
-			name:            "scenario 3: update when desired are not the same as expected RBAC Roles",
-			projectToSync:   test.CreateProject("thunderball", test.CreateUser("James Bond")),
-			expectedActions: []string{"update", "update"},
-			existingClusterRoles: []*rbacv1.ClusterRole{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:owners-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update", "delete"},
-						},
-					},
-				},
-
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update", "delete"},
-						},
-					},
-				},
-
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:viewers-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get"},
-						},
-					},
-				},
-			},
-			expectedClusterRoles: []*rbacv1.ClusterRole{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "kubermatic:project-thunderball:editors-thunderball",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								APIVersion: kubermaticv1.SchemeGroupVersion.String(),
-								Kind:       kubermaticv1.ProjectKindName,
-								Name:       "thunderball",
-								UID:        "thunderballID", // set manually
-							},
-						},
-					},
-					Rules: []rbacv1.PolicyRule{
-						{
-							APIGroups:     []string{kubermaticv1.SchemeGroupVersion.Group},
-							Resources:     []string{"projects"},
-							ResourceNames: []string{"thunderball"},
-							Verbs:         []string{"get", "update"},
-						},
-					},
-				},
-			},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			// setup the test scenario
-			objs := []runtime.Object{}
-			clusterRoleIndexer := cache.NewIndexer(cache.MetaNamespaceKeyFunc, cache.Indexers{})
-			for _, existingClusterRole := range test.existingClusterRoles {
-				err := clusterRoleIndexer.Add(existingClusterRole)
-				if err != nil {
-					t.Fatal(err)
-				}
-				objs = append(objs, existingClusterRole)
-			}
-			fakeKubeClient := fake.NewSimpleClientset(objs...)
-			clusterRoleLister := rbaclister.NewClusterRoleLister(clusterRoleIndexer)
-
-			// act
-			target := Controller{}
-			err := target.ensureClusterRBACRoleForNamedResource(test.projectToSync.Name, kubermaticv1.ProjectResourceName, kubermaticv1.ProjectKindName, test.projectToSync.GetObjectMeta(), fakeKubeClient, clusterRoleLister)
-
-			// validate
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			if len(test.expectedClusterRoles) == 0 {
-				if len(fakeKubeClient.Actions()) != 0 {
-					t.Fatalf("unexpected actions %#v", fakeKubeClient.Actions())
-				}
-				return
-			}
-
-			if len(fakeKubeClient.Actions()) != len(test.expectedClusterRoles) {
-				t.Fatalf("unexpected actions %#v ", fakeKubeClient.Actions())
-			}
-
-			for index, action := range fakeKubeClient.Actions() {
-				if !action.Matches(test.expectedActions[index], "clusterroles") {
-					t.Fatalf("unexpected action %#v", action)
-				}
-				// TODO: figure out why action.(clienttesting.GenericAction) does not work
-				createAction, ok := action.(clienttesting.CreateAction)
-				if !ok {
-					t.Fatalf("unexpected action %#v", action)
-				}
-				if !equality.Semantic.DeepEqual(createAction.GetObject().(*rbacv1.ClusterRole), test.expectedClusterRoles[index]) {
-					t.Fatalf("%v", diff.ObjectDiff(test.expectedClusterRoles[index], createAction.GetObject().(*rbacv1.ClusterRole)))
-				}
-			}
-		})
-	}
-}
-
 func TestEnsureProjectOwner(t *testing.T) {
 	tests := []struct {
 		name            string
@@ -2088,7 +1423,7 @@ func TestEnsureProjectOwner(t *testing.T) {
 			bindingLister := kubermaticv1lister.NewUserProjectBindingLister(bindingIndexer)
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.userLister = userLister
 			target.userProjectBindingLister = bindingLister
@@ -2301,7 +1636,7 @@ func TestEnsureProjectRBACRoleForResources(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.projectResources = test.projectResourcesToSync
 			target.seedClusterProviders = seedClusterProviders
@@ -2644,7 +1979,7 @@ func TestEnsureProjectRBACRoleBindingForResources(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.projectResources = test.projectResourcesToSync
 			target.seedClusterProviders = seedClusterProviders
@@ -2911,7 +2246,7 @@ func TestEnsureProjectCleanUpForRoleBindings(t *testing.T) {
 			}
 
 			// act
-			target := Controller{}
+			target := projectController{}
 			target.masterClusterProvider = fakeMasterClusterProvider
 			target.projectResources = test.projectResourcesToSync
 			target.seedClusterProviders = seedClusterProviders
